+++ internal/auth/middelware.go
@@ -1,0 +1,16 @@
+package auth
+
+import (
+	echojwt "github.com/labstack/echo-jwt/v4"
+	"github.com/labstack/echo/v4"
+)
+
+// CreateJwtMiddleware creates a middleware function for JWTs
+// TODO expand the config settings
+func CreateJwtMiddleware(secret []byte) echo.MiddlewareFunc {
+	config := echojwt.Config{
+		SigningKey: secret,
+	}
+
+	return echojwt.WithConfig(config)
+}
